Version Numbers: Avoid "inactive" #s by slurping from enter_bug.cgi
This commit refactors the code in query.pl that grabs version #s to
allow us to pull version numbers from either the query.cgi or the
enter_bug.cgi page on Bugzilla, defaulting to the latter.

Because we have marked several LibreOffice version #s as "inactive,"
these version #s do not appear on enter_bug.cgi (in a drop-down box)
but *do* appear on the query.cgi page.

This commit is a quick and dirty way to grab only the "active"
versions. There's hopefully a cleaner way to grab this list from
Bugzilla...